move verify_paths to debug commands
